AbstractPotentials

Functions

FreeBird.AbstractPotentials — Module

AbstractPotentials

Module for defining and implementing potentials.

FreeBird.AbstractPotentials.CompositeParameterSets — Type

struct CompositeParameterSets{C,P} <: MultiComponentParameterSets

CompositeParameterSets is a struct that represents a set of composite parameter sets, typically used for multi-component systems.

Fields

param_sets::Matrix{P}: A matrix of parameter sets representing the composite parameter sets.

Type Parameters

C::Int: The number of composite parameter sets.

FreeBird.AbstractPotentials.CompositeParameterSets — Method

CompositeParameterSets(c::Int, ps::Vector{P}) where {P <: SingleComponentParameterSets}

Construct a CompositeParameterSets object from a vector of SingleComponentParameterSets.

Arguments

c::Int: The number of components.

ps::Vector{LJParameters}: A vector of LJParameters.

The number of elements in the vector must be equal to c^2 or c*(c+1)/2. The former case is for a full flattened matrix of LJParameters, useful when the interactions are asymmetric, i.e., epsilon_ij != epsilon_ji. The latter case is for symmetric interactions, i.e., epsilon_ij = epsilon_ji, hence only the upper triangular part of the matrix is needed.

Returns

A CompositeParameterSets object.

Example

julia> ps = [LJParameters(epsilon=e) for e in [11, 12, 13, 22, 23, 33]]
6-element Vector{LJParameters}:
 LJParameters(11.0 eV, 1.0 Å, Inf, 0.0 eV)
 LJParameters(12.0 eV, 1.0 Å, Inf, 0.0 eV)
 LJParameters(13.0 eV, 1.0 Å, Inf, 0.0 eV)
 LJParameters(22.0 eV, 1.0 Å, Inf, 0.0 eV)
 LJParameters(23.0 eV, 1.0 Å, Inf, 0.0 eV)
 LJParameters(33.0 eV, 1.0 Å, Inf, 0.0 eV)

julia> ljp = CompositeParameterSets(3,ps)
CompositeParameterSets{3}(param_sets::3x3 Matrix{LJParameters}):
    param_sets[1, 1] : LJParameters(11.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[1, 2] : LJParameters(12.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[1, 3] : LJParameters(13.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[2, 1] : LJParameters(12.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[2, 2] : LJParameters(22.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[2, 3] : LJParameters(23.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[3, 1] : LJParameters(13.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[3, 2] : LJParameters(23.0 eV, 1.0 Å, Inf, 0.0 eV)
    param_sets[3, 3] : LJParameters(33.0 eV, 1.0 Å, Inf, 0.0 eV)

FreeBird.AbstractPotentials.GuptaParameters — Type

struct GuptaParameters

The GuptaParameters struct represents the parameters for the Gupta potential.

Fields

A::typeof(1.0u"eV"): The repulsive energy scale of the potential.

ξ::typeof(1.0u"eV"): The attractive energy scale of the potential.

p::Float64: The exponent for the repulsive term.

q::Float64: The exponent for the attractive term.

r0::typeof(1.0u"Å"): The equilibrium distance for the potential.

cutoff::Float64: The cutoff distance for the potential, in units of r0.

FreeBird.AbstractPotentials.GuptaParameters — Method

GuptaParameters(;A=1.0, ξ=1.0, p=10.0, q=5.0, r0=1.0, cutoff=Inf)

A constructor for the GuptaParameters struct with default values for the Gupta potential with no cutoff.

FreeBird.AbstractPotentials.LJParameters — Type

struct LJParameters

The LJParameters struct represents the parameters for the Lennard-Jones potential.

Fields

epsilon::typeof(1.0u"eV"): The energy scale of the potential.

sigma::typeof(1.0u"Å"): The length scale of the potential.

cutoff::Float64: The cutoff distance for the potential, in units of sigma.

shift::typeof(0.0u"eV"): The energy shift applied to the potential, calculated at the cutoff distance.

FreeBird.AbstractPotentials.LJParameters — Method

LJParameters(;epsilon=1.0, sigma=1.0, cutoff=Inf, shift=true)

A constructor for the LJParameters struct with default values for the Lennard-Jones potential with no cutoff or shift. The shift parameter can be specified as a boolean, if true, the shift energy is calculated automatically at the cutoff distance; or as a typeof(0.0u"eV"), in which case the value is used directly.

Example

julia> lj = LJParameters(epsilon=0.1,sigma=2.5,cutoff=3.5,shift=false)
LJParameters(0.1 eV, 2.5 Å, 3.5, 0.0 eV)

julia> lj = LJParameters(sigma=2.5)
LJParameters(1.0 eV, 2.5 Å, Inf, 0.0 eV)

julia> lj = LJParameters(cutoff=3.5,shift=5.0)
LJParameters(1.0 eV, 1.0 Å, 3.5, 5.0 eV)

julia> lj = LJParameters(cutoff=3.5,shift=true)
LJParameters(1.0 eV, 1.0 Å, 3.5, -0.0021747803916549904 eV)

julia> lj = LJParameters(cutoff=3.5,shift=false)
LJParameters(1.0 eV, 1.0 Å, 3.5, 0.0 eV)

FreeBird.AbstractPotentials.lj_energy — Method

lj_energy(r::typeof(1.0u"Å"), lj::LJParameters)

Compute the Lennard-Jones energy between two particles at a given distance.

Arguments

r::typeof(1.0u"Å"): The distance between the particles.

lj::LJParameters: The Lennard-Jones parameters.

Returns

0.0u"eV" if the distance is greater than the cutoff distance.

The Lennard-Jones energy minus the shift otherwise.

FreeBird.AbstractPotentials.lj_energy — Method

lj_energy(epsilon::typeof(1.0u"eV"), sigma::typeof(1.0u"Å"), r::typeof(1.0u"Å"))

Compute the Lennard-Jones potential energy between two particles.

The Lennard-Jones potential energy is given by the equation:

\[V(r_{ij}) = 4\varepsilon_{ij} \left[\left(\frac{\sigma_{ij}}{r_{ij}}\right)^{12} - \left(\frac{\sigma_{ij}}{r_{ij}}\right)^{6}\right]\]

where epsilon is the energy scale, sigma is the distance scale, and r is the distance between the particles.

Arguments

epsilon::typeof(1.0u"eV"): The energy scale of the potential.

sigma::typeof(1.0u"Å"): The distance scale of the potential.

r::typeof(1.0u"Å"): The distance between the particles.

Returns

The Lennard-Jones potential energy between the particles.

FreeBird.AbstractPotentials.pair_energy — Method

pair_energy(r::typeof(1.0u"Å"), lj::LJParameters)

Compute the energy of a pair of particles separated by distance r using the Lennard-Jones potential.

Arguments

r::typeof(1.0u"Å"): The distance between the two particles.

lj::LJParameters: The Lennard-Jones parameters.

Returns

energy::typeof(0.0u"eV"): The energy of the pair of particles.