EnergyEval

Functions

FreeBird.EnergyEval — Module

Module for evaluating energy-related quantities for a system.

FreeBird.EnergyEval.free_free_energy — Method

free_free_energy(at::AbstractSystem, lj::LJParameters; frozen::Int64=0)

Calculate the energy from interactions between free particles using the Lennard-Jones potential. The energy is calculated by summing the pairwise interactions between the free particles.

Arguments

at::AbstractSystem: The system for which the energy is calculated.

lj::LJParameters: The Lennard-Jones parameters.

frozen::Int64: The number of frozen particles in the system.

Returns

free_free_energy: The energy from interactions between free particles.

FreeBird.EnergyEval.free_frozen_energy — Method

free_frozen_energy(at::AbstractSystem, lj::LJParameters, frozen::Int64)

Compute the energy from interactions between free and frozen particles using the Lennard-Jones potential. The energy is calculated by summing the pairwise interactions between the free and frozen particles.

Arguments

at::AbstractSystem: The system containing the particles.

lj::LJParameters: The Lennard-Jones parameters.

frozen::Int64: The number of frozen particles.

Returns

free_frozen_energy: The energy from interactions between free and frozen particles.

FreeBird.EnergyEval.frozen_energy — Method

frozen_energy(at::AbstractSystem, lj::LJParameters, frozen::Int64)

Compute the energy of the frozen particles in the system using the Lennard-Jones potential. The energy is calculated by summing the pairwise interactions between the frozen particles. Since the frozen particles do not move, the energy is typically only calculated once for a given system.

Arguments

at::AbstractSystem: The system containing the particles.

lj::LJParameters: The Lennard-Jones parameters.

frozen::Int64: The number of frozen particles.

Returns

frozen_energy: The energy of the frozen particles in the system.

FreeBird.EnergyEval.interaction_energy — Method

interaction_energy(at::AbstractSystem, lj::LJParameters; frozen::Int64=0)

Compute the energy from interactions between particles using the Lennard-Jones potential. The energy is calculated by combining the energies from interactions between free particles and between free and frozen particles. The energy from interactions between frozen particles is not included, as it is typically treated as a constant energy offset. If the number of frozen particles is zero, the energy calculated is equivalent to the total energy of the system.

Arguments

at::AbstractSystem: The system containing the particles.

lj::LJParameters: The Lennard-Jones parameters.

frozen::Int64: The number of frozen particles.

Returns

interaction_energy: The energy from interactions between particles.

FreeBird.EnergyEval.pbc_dist — Method

pbc_dist(pos1, pos2, at)

Compute the distance between two positions considering periodic boundary conditions. Currently only works for orthorhombic lattices.

Arguments

pos1::Union{SVector{T},Vector{T}}: The first position.

pos2::Union{SVector{T},Vector{T}}: The second position.

at::AbstractSystem: The abstract system containing boundary conditions and bounding box.

Returns

d::Float64: The distance between pos1 and pos2 considering periodic boundary conditions.