Monte Carlo Moves

Functions

FreeBird.MonteCarloMoves — Module

Module containing functions for performing Monte Carlo moves on atomic/molecular systems.

FreeBird.MonteCarloMoves.MC_new_sample! — Method

MC_new_sample!(lattice::LatticeWalker, h::ClassicalHamiltonian, emax::Float64; energy_perturb::Float64=0.0)

Generate a new sample for the lattice system.

Arguments

lattice::LatticeWalker: The walker to generate a new sample for.

h::ClassicalHamiltonian: The Hamiltonian containing the on-site and nearest-neighbor interaction energies.

emax::Float64: The maximum energy allowed for accepting a move.

energy_perturb::Float64=0.0: The energy perturbation used to make degenerate configurations distinguishable.

Returns

accept_this_walker::Bool: Whether the walker is accepted or not.

lattice::LatticeWalker: The updated walker.

FreeBird.MonteCarloMoves.MC_nve_walk! — Method

MC_nve_walk!(
    n_steps::Int, 
    at::AtomWalker{C}, 
    lj::LennardJonesParametersSets, 
    step_size::Float64; 
    e_demon_tolerance=1e-9u"eV",
    demon_energy_threshold=Inf*u"eV",
    demon_gain_threshold=Inf*u"eV",
    max_add_steps::Int=1_000_000
)

Perform a NVE walk using the demon algorithm. The demon algorithm is used to maintain the energy (E) of the system.

Arguments

n_steps::Int: The number of demon walks to perform.

at::AtomWalker: The walker to perform the demon walk on.

lj::LJParameters: The Lennard-Jones parameters for the system.

step_size::Float64: The step size for the demon walk.

Optional Arguments

e_demon_tolerance=1e-9u"eV": The energy tolerance for the demon, below which the demon walk is considered successful, i.e., the NVE condition is satisfied.

demon_energy_threshold=Inf*u"eV": The maximum energy allowed for the demon to have.

demon_gain_threshold=Inf*u"eV": The energy gain threshold for the demon during each move.

max_add_steps::Int=1_000_000: The maximum number of additional demon walks if the demon energy is above the tolerance.

Returns

accept_this_walker::Bool: Whether the walker is accepted or not.

accept_ratio::Float64: The acceptance ratio of the demon walk (excluding additional demon walks).

at_final::AtomWalker: The final walker after the demon walks.

demon_energies::Array{Float64}: The energies of the demon at each step.

temp_estimate::Float64: The estimated temperature of the system.

FreeBird.MonteCarloMoves.MC_random_walk! — Method

MC_random_walk!(n_steps::Int, lattice::LatticeWalker, h::LatticeGasHamiltonian, emax::Float64; energy_perturb::Float64=0.0)

Perform a Monte Carlo random walk on the lattice system.

Arguments

n_steps::Int: The number of Monte Carlo steps to perform.

lattice::LatticeWalker: The walker to perform the random walk on.

h::LatticeGasHamiltonian: The lattice gas Hamiltonian.

emax::Float64: The maximum energy allowed for accepting a move.

energy_perturb::Float64=0.0: The energy perturbation used to make degenerate configurations distinguishable.

Returns

accept_this_walker::Bool: Whether the walker is accepted or not.

accept_rate::Float64: The acceptance rate of the random walk.

lattice::LatticeWalker: The updated walker.

FreeBird.MonteCarloMoves.MC_random_walk! — Method

MC_random_walk!(n_steps::Int, at::AtomWalker, lj::LJParameters, step_size::Float64, emax::typeof(0.0u"eV"))

Perform a Monte Carlo random walk on the atomic/molecular system.

Arguments

n_steps::Int: The number of Monte Carlo steps to perform.

at::AtomWalker{C}: The walker to perform the random walk on.

lj::LennardJonesParametersSets: The Lennard-Jones potential parameters.

step_size::Float64: The maximum distance an atom can move in any direction.

emax::typeof(0.0u"eV"): The maximum energy allowed for accepting a move.

Returns

accept_this_walker::Bool: Whether the walker is accepted or not.

accept_rate::Float64: The acceptance rate of the random walk.

at::AtomWalker: The updated walker.

FreeBird.MonteCarloMoves.additional_demon_walk! — Method

additional_demon_walk!(e_demon::typeof(0.0u"eV"), at::AtomWalker, lj::LJParameters, step_size::Float64;
                      e_demon_tolerance=1e-9u"eV", max_add_steps::Int=1_000_000)

Performs additional demon walk steps until the demon energy e_demon is below the tolerance e_demon_tolerance or the maximum number of additional steps max_add_steps is reached.

Arguments

e_demon::typeof(0.0u"eV"): The initial demon energy.

at::AtomWalker{C}: The walker that the demon walk is performed on.

lj::LennardJonesParametersSets: The LJ parameters.

step_size::Float64: The step size for the demon walk.

e_demon_tolerance=1e-9u"eV": The tolerance for the demon energy.

max_add_steps::Int=1_000_000: The maximum number of additional steps.

Returns

accept_this_walker::Bool: Whether the walker is accepted after the additional demon walk.

accept_rate::Float64: The acceptance rate of the additional demon walk.

at::AtomWalker: The updated walker.

e_demon::typeof(0.0u"eV"): The final demon energy.

FreeBird.MonteCarloMoves.free_par_index — Method

free_par_index(at::AtomWalker{C}) where C

Get the indices of the free particles in the AtomWalker.

FreeBird.MonteCarloMoves.generate_random_new_lattice_sample! — Method

generate_random_new_lattice_sample!(lattice::MLattice{C}) where C

Generate a new random sample for the multi-component lattice system.

Arguments

lattice::MLattice{C}: The lattice system to generate a new sample for.

Returns

lattice::MLattice{C}: The updated lattice system.

FreeBird.MonteCarloMoves.mean_sq_displacement — Method

mean_sq_displacement(at::AtomWalker, at_orig::AtomWalker)

Calculate the mean squared displacement before and after random walk(s). Note that due to the current implementation of the periodic boundary wrap, this function is not appropriate to use for calculating mean displacements in a propagation.

Arguments

at::AtomWalker{C}: The current AtomWalker after the random walk.

at_orig::AtomWalker{C}: The original AtomWalker before the random walk.

Returns

distsq::typeof(0.0u"Å"^2): The mean squared displacement of all free particles.

FreeBird.MonteCarloMoves.periodic_boundary_wrap! — Method

periodic_boundary_wrap!(pos::SVector{3,T}, system::AbstractSystem) where T

Wrap the position vector pos according to the periodic boundary conditions of the system. If the boundary condition is Periodic(), the position is wrapped using the modulo operator. If the boundary condition is DirichletZero(), the position is wrapped by reflecting the position vector across the boundary.

Arguments

pos::SVector{3,T}: The position vector to be wrapped.

system::AbstractSystem: The system containing the periodic boundary conditions.

Returns

The wrapped position vector.

FreeBird.MonteCarloMoves.single_atom_demon_walk! — Method

single_atom_demon_walk!(at::AtomWalker{C}, 
                        lj::LennardJonesParametersSets, 
                        step_size::Float64;
                        e_demon=0.0u"eV",
                        demon_energy_threshold=Inf*u"eV",
                        demon_gain_threshold=Inf*u"eV")

Perform a single atom demon walk.

Arguments:

at::AtomWalker: The walker to perform the demon walk on.

lj::LJParameters: The LJ parameters.

step_size::Float64: The step size for the random walk.

Keyword Arguments:

e_demon=0.0u"eV": The energy of the demon.

demon_energy_threshold=Inf*u"eV": The energy threshold for the demon.

demon_gain_threshold=Inf*u"eV": The energy gain threshold for the demon during each move.

Returns:

accept::Bool: Whether the move is accepted or rejected.

at::AtomWalker: The updated atom walker object.

e_demon::Float64: The updated energy of the demon.

FreeBird.MonteCarloMoves.single_atom_random_walk! — Method

single_atom_random_walk!(pos::SVector{3,T}, step_size::Float64) where T

Perform a single atom random walk by updating the position pos in each direction by a random amount. The step_size determines the maximum distance the atom can move in any direction.

Arguments

pos::SVector{3,T}: The current position of the atom as a 3D vector.

step_size::Float64: The maximum distance the atom can move in any direction.

Returns

pos: The updated position of the atom.