Sampling Schemes

Functions

FreeBird.SamplingSchemes — Module

SamplingSchemes

Module for defining the sampling schemes.

FreeBird.SamplingSchemes.MCNewSample — Type

struct MCNewSample <: MCRoutine

A type for generating a new walker from a random configuration. Currently, it is intended to use this routine for lattice gas systems.

FreeBird.SamplingSchemes.MCRandomWalkClone — Type

struct MCRandomWalkClone <: MCRoutine

A type for generating a new walker by cloning an existing walker and performing a random walk for decorrelation.

FreeBird.SamplingSchemes.MCRandomWalkMaxE — Type

struct MCRandomWalkMaxE <: MCRoutine

A type for generating a new walker by performing a random walk for decorrelation on the highest-energy walker.

FreeBird.SamplingSchemes.MCRoutine — Type

abstract type MCRoutine

An abstract type representing a Monte Carlo routine.

FreeBird.SamplingSchemes.NestedSamplingParameters — Type

mutable struct NestedSamplingParameters <: SamplingParameters

The NestedSamplingParameters struct represents the parameters used in the nested sampling scheme.

Fields

mc_steps::Int64: The number of total Monte Carlo moves to perform.

initial_step_size::Float64: The initial step size, which is the fallback step size if MC routine fails to accept a move.

step_size::Float64: The on-the-fly step size used in the sampling process.

step_size_lo::Float64: The lower bound of the step size.

step_size_up::Float64: The upper bound of the step size.

fail_count::Int64: The number of failed MC moves in a row.

allowed_fail_count::Int64: The maximum number of failed MC moves allowed before resetting the step size.

FreeBird.SamplingSchemes.adjust_step_size — Method

adjust_step_size(ns_params::NestedSamplingParameters, rate::Float64)

Adjusts the step size of the nested sampling algorithm based on the acceptance rate. If the acceptance rate is greater than 0.75, the step size is increased by 1%. If the acceptance rate is less than 0.25, the step size is decreased by 1%.

Arguments

ns_params::NestedSamplingParameters: The parameters of the nested sampling algorithm.

rate::Float64: The acceptance rate of the algorithm.

range::Tuple{Float64, Float64}: The range of acceptance rates for adjusting the step size. Default is (0.25, 0.75).

Returns

ns_params::NestedSamplingParameters: The updated parameters with adjusted step size.

FreeBird.SamplingSchemes.exact_enumeration — Method

exact_enumeration(lattice::SLattice{G}, cutoff_radii::Tuple{Float64, Float64}, h::ClassicalHamiltonian) where G

Enumerate all possible configurations of a lattice system and compute the energy of each configuration.

Arguments

lattice::SLattice{G}: The (starting) lattice system to enumerate. All possible configurations will be generated from this lattice system.

h::ClassicalHamiltonian: The Hamiltonian containing the on-site and nearest-neighbor interaction energies.

Returns

DataFrame: A DataFrame containing the energy and configuration of each configuration.

LatticeGasWalkers: A collection of lattice walkers for each configuration.

FreeBird.SamplingSchemes.nested_sampling_loop! — Method

nested_sampling_loop!(liveset::AtomWalkers, ns_params::NestedSamplingParameters, n_steps::Int64, mc_routine::MCRoutine; args...)

Perform a nested sampling loop for a given number of steps.

Arguments

liveset::AtomWalkers: The initial set of walkers.

ns_params::NestedSamplingParameters: The parameters for nested sampling.

n_steps::Int64: The number of steps to perform.

mc_routine::MCRoutine: The Monte Carlo routine to use.

Returns

df: A DataFrame containing the iteration number and maximum energy for each step.

liveset: The updated set of walkers.

ns_params: The updated nested sampling parameters.

FreeBird.SamplingSchemes.nested_sampling_loop! — Method

nested_sampling_loop!(liveset::LatticeGasWalkers, ns_params::LatticeNestedSamplingParameters, n_steps::Int64, mc_routine::MCRoutine, save_strategy::DataSavingStrategy)

Perform a nested sampling loop on a lattice gas system for a given number of steps.

Arguments

liveset::LatticeGasWalkers: The initial set of walkers.

ns_params::LatticeNestedSamplingParameters: The parameters for nested sampling.

n_steps::Int64: The number of steps to perform.

mc_routine::MCRoutine: The Monte Carlo routine to use.

save_strategy::DataSavingStrategy: The strategy for saving data.

Returns

df: A DataFrame containing the iteration number and maximum energy for each step.

liveset: The updated set of walkers.

ns_params: The updated nested sampling parameters.

FreeBird.SamplingSchemes.nested_sampling_step! — Method

nested_sampling_step!(liveset::AtomWalkers, ns_params::NestedSamplingParameters, mc_routine::MCRandomWalk)

Perform a single step of the nested sampling algorithm using the Monte Carlo random walk routine.

Arguments

liveset::AtomWalkers: The set of atom walkers.

ns_params::NestedSamplingParameters: The parameters for nested sampling.

mc_routine::MCRandomWalk: The Monte Carlo random walk routine. Currently within this function, the random walk is only applied to the highest-energy walker, i.e., the one being culled.

Returns

iter: The iteration number after the step.

emax: The highest energy recorded during the step.

liveset: The updated set of atom walkers.

ns_params: The updated nested sampling parameters.

FreeBird.SamplingSchemes.nested_sampling_step! — Method

nested_sampling_step!(liveset::LatticeGasWalkers, ns_params::LatticeNestedSamplingParameters, mc_routine::MCNewSample; accept_same_config::Bool=false)

Perform a single step of the nested sampling algorithm.

This function takes a liveset of lattice gas walkers, ns_params containing the parameters for nested sampling, and mc_routine representing the Monte Carlo routine for generating new samples. It performs a single step of the nested sampling algorithm by updating the liveset with a new walker.

Arguments

liveset::LatticeGasWalkers: The liveset of lattice gas walkers.

ns_params::LatticeNestedSamplingParameters: The parameters for nested sampling.

mc_routine::MCNewSample: The Monte Carlo routine for generating new samples.

accept_same_config::Bool=true: A flag indicating whether to accept a new sample with the same configuration as an existing one.

Returns

iter: The iteration number of the liveset after the step.

emax: The maximum energy of the liveset after the step.

liveset::LatticeGasWalkers: The updated liveset after the step.

ns_params::LatticeNestedSamplingParameters: The updated nested sampling parameters after the step.

FreeBird.SamplingSchemes.nested_sampling_step! — Method

nested_sampling_step!(liveset::LatticeGasWalkers, ns_params::LatticeNestedSamplingParameters, mc_routine::MCRoutine; accept_same_config::Bool=true)

Perform a single step of the nested sampling algorithm.

This function takes a liveset of lattice gas walkers, ns_params containing the parameters for nested sampling, and mc_routine representing the Monte Carlo routine for generating new samples. It performs a single step of the nested sampling algorithm by updating the liveset with a new walker.

Arguments

liveset::LatticeGasWalkers: The liveset of lattice gas walkers.

ns_params::LatticeNestedSamplingParameters: The parameters for nested sampling.

mc_routine::MCRoutine: The Monte Carlo routine for generating new samples.

Returns

iter: The iteration number of the liveset after the step.

emax: The maximum energy of the liveset after the step.

FreeBird.SamplingSchemes.sort_by_energy! — Method

sort_by_energy!(liveset::LJAtomWalkers)

Sorts the walkers in the liveset by their energy in descending order.

Arguments

liveset::LJAtomWalkers: The liveset of walkers to be sorted.

Returns

liveset::LJAtomWalkers: The sorted liveset.

FreeBird.SamplingSchemes.update_iter! — Method

update_iter!(liveset::AtomWalkers)

Update the iteration count for each walker in the liveset.

Arguments

liveset::AtomWalkers: The set of walkers to update.